+++ CHANGELOG.md
@@ -3,9 +3,16 @@
 # Version 0.4.0
 
 - **Updated to Bevy 0.16.0**
-- Added `no_std` support
+- **Extended reaction components:**
+    - Removed `Single` variant from `DespawnOnExit` and `DespawnOnDisable` components
+    - Renamed `DespawnOnExit` -> `DespawnOnExitState` component
+    - Renamed `DespawnOnDisable` -> `DespawnOnDisableState` component
+    - Renamed `VisibleWhileIn` -> `VisibleInState` component
+    - Renamed `VisibleWhileEnabled` -> `VisibleInEnabledState` component
+    - Added `EnabledInState` component
+    - Added `EnabledInEnabledState` component
+- **Added `no_std` support**
 - Added `NextStateStackCommandsExt` extension trait
-- Removed `Single` variant from `DespawnOnExit` and `DespawnOnDisable` components
 
 # Version 0.3.0